roblox-sniping-bot/requirements.txt
requests==2.26.0
pandas==1.5.3
scikit-learn==1.2.2
joblib==1.1.1
websocket-client==1.2.1
python-dotenv==0.20.0
